[preprocessor/folder] add post to webhook client
